+++ .pylintrc
@@ -433,7 +433,10 @@ good-names=i,
            _,
            x,
            y,
-           lr
+           lr,
+           n,
+           t,
+           e
 
 # Include a hint for the correct naming format with invalid-name.
 include-naming-hint=no
