+++ all/memory/n_step_test.py
@@ -3,6 +3,7 @@ import torch
 import torch_testing as tt
 from all.memory import NStepBuffer
 
+
 class NStepBufferTest(unittest.TestCase):
     def test_rollout(self):
         buffer = NStepBuffer(2, 3, discount_factor=0.5)
@@ -62,10 +63,10 @@ class NStepBufferTest(unittest.TestCase):
 
     def test_multi_rollout(self):
         buffer = NStepBuffer(2, 2, discount_factor=0.5)
-        s = ['state' + str(i) for i in range(12)]
-        buffer.store(s[0:2], torch.ones(2))
-        buffer.store(s[2:4], torch.ones(2))
-        buffer.store(s[4:6], torch.ones(2))
+        raw_states = ['state' + str(i) for i in range(12)]
+        buffer.store(raw_states[0:2], torch.ones(2))
+        buffer.store(raw_states[2:4], torch.ones(2))
+        buffer.store(raw_states[4:6], torch.ones(2))
 
         states, next_states, returns = buffer.sample(-1)
         self.assert_array_equal(states, ['state' + str(i) for i in range(4)])
@@ -74,20 +75,22 @@ class NStepBufferTest(unittest.TestCase):
         ])
         tt.assert_allclose(returns, torch.tensor([1.5, 1.5, 1, 1]))
 
-        buffer.store(s[6:8], torch.ones(2))
-        buffer.store(s[8:10], torch.ones(2))
+        buffer.store(raw_states[6:8], torch.ones(2))
+        buffer.store(raw_states[8:10], torch.ones(2))
 
         states, next_states, returns = buffer.sample(-1)
-        self.assert_array_equal(states, ['state' + str(i) for i in range(4, 8)])
+        self.assert_array_equal(
+            states, ['state' + str(i) for i in range(4, 8)])
         self.assert_array_equal(next_states, [
             'state8', 'state9', 'state8', 'state9'
         ])
         tt.assert_allclose(returns, torch.tensor([1.5, 1.5, 1, 1]))
 
-
     def assert_array_equal(self, actual, expected):
         for i, exp in enumerate(expected):
-            self.assertEqual(actual[i], exp, msg=(("\nactual: %s\nexpected: %s") % (actual, expected)))
+            self.assertEqual(actual[i], exp, msg=(
+                ("\nactual: %s\nexpected: %s") % (actual, expected)))
+
 
 if __name__ == '__main__':
     unittest.main()
